re-org the class folders for better dependency managerment
