+++ orttraining/orttraining/core/framework/checkpoint_common.cc
@@ -3,10 +3,7 @@
 
 #include "core/common/logging/logging.h"
 #include "core/common/logging/sinks/clog_sink.h"
-#include "core/common/path.h"
-#include "core/common/path_string.h"
 #include "core/common/status.h"
-#include "core/framework/data_transfer_manager.h"
 #include "core/framework/data_types.h"
 #include "core/framework/framework_common.h"
 #include "core/framework/tensorprotoutils.h"
